% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{ana_sumVol}
\alias{ana_sumVol}
\title{Summary label/annotation records for volume}
\usage{
ana_sumVol(
  volName,
  countLimit = 0,
  haveReport,
  url = NULL,
  auth = NULL,
  downloadFolder = NULL,
  waitingTime = 5,
  labelReport = NULL,
  annotationReport = NULL
)
}
\arguments{
\item{volName}{Volume name.}

\item{countLimit}{(OPTIONAL) Exclude the label/annotation records that count less than the limit (int). Default: 0.}

\item{haveReport}{TRUE or FALSE. Already have the reports or not.\cr
If FALSE, \cr input need input parameters: \code{url}, \code{auth}, \code{downloadFolder}, \code{waitingTime}; it can automatically call \code{\link{reqNgetRep}} to prepare volume's reports. \cr
If TRUE, \cr input the report: \code{labelReport} and \code{annotationReport}.}

\item{url}{BIIGLE server URL (e.g. "https://seaimage.oc.ntu.edu.tw").}

\item{auth}{Authentication string generated by \code{\link{genAuth}}.}

\item{downloadFolder}{The directory path to save the downloaded report files.}

\item{waitingTime}{(OPTIONAL) Waiting time (in sec.)for server to generate the report. Default: 5 sec.}

\item{labelReport}{The label report table.}

\item{annotationReport}{The annotation report table.}
}
\description{
Visualize label/annotation count as a bar chart in this volume.
}
\examples{
\dontrun{
           countLimit = 20,
ana_sumVol(volName = "TTC19",
           countLimit = 20,
           haveReport = FALSE,
           url = "https://seaimage.oc.ntu.edu.tw",
           auth = myAuth,
           downloadFolder = "C:\\\\tmp\\\\odbiigleT\\\\tmp",
           waitingTime = 5)
}
}
