% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBIIGLE.R
\docType{package}
\name{biigler}
\alias{biigler}
\title{biigler}
\description{
\code{biigler} is an unofficial R package developed by Underwater Image Identification Working Group in ODB, IONTU, mainly for substituting the repetitive operations on the BIIGLE 2.0 graphical user interface.
It is based on BIIGLE v3.14.2 and R v4.1.2.
}
