% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volPostNDelete.R
\name{delAllAnnos}
\alias{delAllAnnos}
\title{Delete all annotations on images in filename order.}
\usage{
delAllAnnos(url, auth, volName, imgStart = NULL, imgEnd = NULL)
}
\arguments{
\item{url}{BIIGLE server URL (e.g. "https://seaimage.oc.ntu.edu.tw").}

\item{auth}{Authentication string generated by \code{\link{genAuth}}.}

\item{volName}{The volume going to delete annotations.}

\item{imgStart}{(OPTIONAL) The start image's filename. Default: NULL (If \code{imgStart} and \code{imgEnd} are both \code{NULL}, it will delete for ALL annotations in the volume).}

\item{imgEnd}{(OPTIONAL) The end image's filename Default: NULL (If only 1 image to delete, set the same as the \code{imgStart}; If \code{imgStart} and \code{imgEnd} are both \code{NULL}, it will delete for ALL annotations in the volume).}
}
\description{
Delete all annotations on images IN \code{filename} ORDER! Hence, DO NOT use "Random" to sort files on BIIGLE, otherwise you may misleading yourself which image's annotation you are going to delete.
}
\examples{
\dontrun{
delAllAnnos(url = "https://seaimage.oc.ntu.edu.tw",
            auth = myAuth,
            volName = "odbiigle",
            imgStart = "2006_10_07_16_29_30.jpg",
            imgEnd = "2006_10_07_16_29_40.jpg")
}
}
