% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volPostNDelete.R
\name{delAllLabs}
\alias{delAllLabs}
\title{Delete all labels on images in filename order.}
\usage{
delAllLabs(url, auth, volName, imgStart = NULL, imgEnd = NULL)
}
\arguments{
\item{url}{BIIGLE server URL (e.g. "https://seaimage.oc.ntu.edu.tw").}

\item{auth}{Authentication string generated by \code{\link{genAuth}}.}

\item{volName}{The volume going to delete labels.}

\item{imgStart}{(OPTIONAL) The start image's filename. Default: NULL (If \code{imgStart} and \code{imgEnd} are both \code{NULL}, it will delete for ALL labels in the volume).}

\item{imgEnd}{(OPTIONAL) The end image's filename Default: NULL (If only 1 image to delete, set the same as the \code{imgStart}; If \code{imgStart} and \code{imgEnd} are both \code{NULL}, it will delete for ALL labels in the volume).}
}
\description{
Delete all labels on images IN \code{filename} ORDER! Hence, DO NOT use "Random" to sort files on BIIGLE, otherwise you may misleading yourself which image's label you are going to delete.
}
\examples{
\dontrun{
delAllLabs(url = "https://seaimage.oc.ntu.edu.tw",
           auth = myAuth,
           volName = "odbiigle",
           imgStart = "2006_10_07_16_29_30.jpg",
           imgEnd = "2006_10_07_16_29_40.jpg")
}
}
