% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBIIGLE.R
\name{genAuth}
\alias{genAuth}
\title{Generate Authentication String}
\usage{
genAuth(account, apiToken)
}
\arguments{
\item{account}{Your BIIGLE account (email you sign-in for BIIGLE).}

\item{apiToken}{Your BIIGLE API token.}
}
\description{
Generate the string for BIIGLE API authentication. Use it wherever the function needs to interact with BIIGLE API (function with the parameter \code{auth}).\cr
NOTE: You need input the API token which can be created by: BIIGLE server > Settings > Tokens > Generate.
}
\details{
See more details on \href{https://biigle.de/doc/api/index.html}{BIIGLE RestfulAPI manual}.
}
\examples{
\dontrun{
forAuth <- genAuth(account = "myBiigleAccount@email.com",
                   apiToken = "myAPItoken")
}
}
