% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBIIGLE.R
\docType{package}
\name{getBIIGLE}
\alias{getBIIGLE}
\title{Get information on BIIGLE server.}
\description{
Get information on BIIGLE server, including basic volume ID, label tree information, and so on.
}
\section{Functions}{

\itemize{
\item \code{genAuth} Generate the string for BIIGLE API authentication.\cr
\item \code{getTree} Get all label's information in the label tree through tree ID.\cr
\item \code{getVolID} Get volume ID through volume name.\cr
\item \code{getVolImg} Get all image's information in the volume.\cr
\item \code{reqNgetRep} Request BIIGLE server to generate a volume report (csv type) and download it.\cr
\item \code{readFiles} Read all reports in the folder and bind as one table.\cr
}
}

