% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBIIGLE.R
\name{getTree}
\alias{getTree}
\title{Get Label Tree}
\usage{
getTree(url, auth, treeID)
}
\arguments{
\item{url}{BIIGLE server URL (e.g. "https://seaimage.oc.ntu.edu.tw").}

\item{auth}{Authentication string generated by \code{\link{genAuth}}.}

\item{treeID}{Label tree ID.}
}
\description{
Get all label's information in the label tree through tree ID.
}
\examples{
\dontrun{
catamiTree <- getTree(url = "https://seaimage.oc.ntu.edu.tw",
                      auth = myAuth,
                      treeID = 5)
}
}
