% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBIIGLE.R
\name{getVolImg}
\alias{getVolImg}
\title{Get Image's information in Volume}
\usage{
getVolImg(url, auth, vol)
}
\arguments{
\item{url}{BIIGLE server URL (e.g. "https://seaimage.oc.ntu.edu.tw").}

\item{auth}{Authentication string generated by \code{\link{genAuth}}.}

\item{vol}{Volume ID (integer) or name (character). Volume's ID can get by \code{\link{getVolID}}.}
}
\description{
Get all image's information in the volume (a table with column "imageID" and "filename").
}
\examples{
\dontrun{
imgList_ttc17 <- getVolImg(url = "https://seaimage.oc.ntu.edu.tw",
                           auth = myAuth,
                           vol = 5)
imgList_ttc17 <- getVolImg(url = "https://seaimage.oc.ntu.edu.tw",
                           auth = myAuth,
                           vol = "TTC17")
}
}
