% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelingAssistant.R
\docType{package}
\name{labelingAssistant}
\alias{labelingAssistant}
\title{Labeling Assistant}
\description{
Post massive label combination on images through BIIGLE API.\cr
It achieve with a set of functions, following 3 steps:
\enumerate{
\item \code{setVol} Set working volume. It only need to run 1 times for a R session.
\item \code{setShortcut} Create the label combination (shortcut) to post on images.
\item \code{postLabs} or \code{delLabs} Post/Delete specific label combination (shortcut) on images in filename order.
}
}
