% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volPostNDelete.R
\name{postAnnos_re}
\alias{postAnnos_re}
\title{Post annotations for volume (reconstruct the volume)}
\usage{
postAnnos_re(url, auth, volName, treeIDs, annoReport, checkFilename = TRUE)
}
\arguments{
\item{url}{BIIGLE server URL (e.g. "https://seaimage.oc.ntu.edu.tw").}

\item{auth}{Authentication string generated by \code{\link{genAuth}}.}

\item{volName}{The empty volume going to post (reconstruct) annotations.}

\item{treeIDs}{Label trees  that need to use in this volume (vector with integer).}

\item{annoReport}{The annotation report table (csv type). It can get by \code{\link{reqNgetRep}}.}

\item{checkFilename}{(OPTIONAL) Check whether the filename of input annotation report all exist in volume's image (imgList). Default: TRUE}
}
\description{
Reconstruct the volume's annotation from a annotation report (must be csv type).
For example, there are 2 volume contain the same image files. "Volume A" have some annotation records and "Volume B" is empty. You can use this function and input the annotation report of "Volume A" to post the same annotation record on "Volume B".\cr
Note that: This function post annotations based on the filename. So, the 2 volume must have the same image filename.
}
\examples{
\dontrun{
postAnnos_re(url = "https://seaimage.oc.ntu.edu.tw",
             auth = myAuth,
             volName = "odbiigle",
             treeIDs = 5,
             annoReport = testReport,
             checkFilename = TRUE)
}
}
