% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volPostNDelete.R
\name{postLabs_re}
\alias{postLabs_re}
\title{Post labels for volume (reconstruct the volume)}
\usage{
postLabs_re(url, auth, volName, treeIDs, labReport, checkFilename = TRUE)
}
\arguments{
\item{url}{BIIGLE server URL (e.g. "https://seaimage.oc.ntu.edu.tw").}

\item{auth}{Authentication string generated by \code{\link{genAuth}}.}

\item{volName}{The empty volume going to post (reconstruct) labels.}

\item{treeIDs}{Label trees  that need to use in this volume (vector with integer).}

\item{labReport}{The label report table (csv type). It can get by \code{\link{reqNgetRep}}.}

\item{checkFilename}{(OPTIONAL) Check whether the filename of input label report all exist in volume's image (imgList). Default: TRUE.}
}
\description{
Reconstruct the volume's label from a label report (must be 'csv' type).\cr
For example, there are 2 volume contain the same image files. "Volume A" have some label records and "Volume B" is empty. You can use this function and input the label report of "Volume A" to post the same label record on "Volume B".\cr
Note that: This function post labels based on the filename. So, the 2 volume must have the same image filename.
}
\examples{
\dontrun{
postLabs_re(url = "https://seaimage.oc.ntu.edu.tw",
            auth = myAuth,
            volName = "odbiigle",
            treeIDs = 5,
            labReport = testReport,
            checkFilename = TRUE)
}
}
