% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBIIGLE.R
\name{readFiles}
\alias{readFiles}
\title{Read all reports in folder and bind as one table}
\usage{
readFiles(folderPath)
}
\arguments{
\item{folderPath}{The directory path to read all report files.}
}
\description{
Some volume's report separated by different tree, read all of them as one table.
}
\examples{
\dontrun{
ttc17 <- readFiles(folderPath = "C:\\\\tmp\\\\odbiigleT\\\\5_csv_image_annotation_report")
}
}
