% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBIIGLE.R
\name{reqNgetRep}
\alias{reqNgetRep}
\title{Request and Download BIIGLE Report}
\usage{
reqNgetRep(
  url,
  auth,
  vol,
  reportType,
  downloadFolder,
  folderName = "biigle_report",
  waitingTime = 5
)
}
\arguments{
\item{url}{BIIGLE server URL (e.g. "https://seaimage.oc.ntu.edu.tw").}

\item{auth}{Authentication string generated by \code{\link{genAuth}}.}

\item{vol}{Volume ID (int) or name (char). Volume ID can get by \code{\link{getVolID}}.}

\item{reportType}{Input 'a' for annotation report or 'l' for label report (only return 'csv' type report).}

\item{downloadFolder}{The directory path to save the downloaded report files.}

\item{folderName}{(OPTIONAL) Customize directory name for downloaded report. Default: "biigle_report".}

\item{waitingTime}{(OPTIONAL) Waiting time (in sec.) for server to generate the report. Default: 5 sec.}
}
\description{
Request BIIGLE server to generate a label/annotation report (limited for 'csv' type) and download it.
}
\examples{
\dontrun{
reqNgetRep(url = "https://seaimage.oc.ntu.edu.tw",
            auth = myAuth,
            vol = "test_exam_answer",
            reportType = "a",
            downloadFolder = "C:\\\\tmp\\\\test",
            folderName = "biigle_report",
            waitingTime = 5)
}
}
