% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelingAssistant.R
\name{setShortcut}
\alias{setShortcut}
\title{Set shortcut (label combination)}
\usage{
setShortcut(x)
}
\arguments{
\item{x}{A label name vector.}
}
\description{
A function of "Labeling Assistant".\cr
Create the label combination (shortcut) to post on images.\cr
NOTE: Need to \code{\link{setVol}} first!
}
\examples{
\dontrun{
setVol(url = "https://seaimage.oc.ntu.edu.tw",
       auth = myAuth,
       vol = "TTC17",
       treeIDs = c(5, 8, 9, 14, 15)) # only need to run 1 time
sho1 <- setShortcut(c("Flat", "Sand / mud (<2mm)", "Bioturbated"))
shoDL <- setShortcut("done_labelling")
}
}
