% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelingAssistant.R
\name{setVol}
\alias{setVol}
\title{Set working volume}
\usage{
setVol(url, auth, vol, treeIDs)
}
\arguments{
\item{url}{BIIGLE server URL (e.g. "https://seaimage.oc.ntu.edu.tw").}

\item{auth}{Authentication string generated by \code{\link{genAuth}}.}

\item{vol}{Volume ID (int) or name (char). Volume's ID can get by \code{\link{getVolID}}.}

\item{treeIDs}{Label trees  that need to use in this volume (vector with integer).}
}
\description{
A function of "Labeling Assistant".\cr
It will create two needed tables "allLabs" and "imgList" in global environment.
}
\examples{
\dontrun{
setVol(url = "https://seaimage.oc.ntu.edu.tw",
       auth = myAuth,
       vol = "TTC17",
       treeIDs = c(5, 8, 9, 14, 15))
}
}
