% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volPostNDelete.R
\docType{package}
\name{volPostNDelete}
\alias{volPostNDelete}
\title{Massive Post and Delete Label/Annotation in a volume}
\description{
Functions about posting/deleting massive amounts of label/annotation records in the volume.
}
\section{Functions}{

\itemize{
\item \code{postLabs_re} Reconstruct the volume's label from a label report (must be 'csv' type).
\item \code{postAnnos_re} Reconstruct the volume's annotation from a annotation report (must be 'csv' type).
\item \code{delAllLabs} Delete all labels on images in filename order.
\item \code{delAllAnnos} Delete all annotations on images in filename order.
}
}

